create setting component and fetch data from supabase for setting components

+++ 17-the-wild-oasis/me-the-wild-oasis/src/services/apiCabins.js
@@ -35,6 +35,7 @@ export const createEditCabin = async (newCabin, id) => {
   }
 
   // 2. Upload Image
+  if (hasImagePath) return data;
   const { error: storageError } = await supabase.storage.from("cabin-images").upload(imageName, newCabin.image);
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.sid);
